+++ test/AutoReview/SrcCodeTest.php
@@ -1,0 +1,39 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Copyright (c) 2017 Andreas Möller.
+ *
+ * For the full copyright and license information, please view
+ * the LICENSE file that was distributed with this source code.
+ *
+ * @link https://github.com/localheinz/specification
+ */
+
+namespace Localheinz\Specification\Test\AutoReview;
+
+use Localheinz\Test\Util\Helper;
+use PHPUnit\Framework;
+
+/**
+ * @internal
+ */
+final class SrcCodeTest extends Framework\TestCase
+{
+    use Helper;
+
+    public function testSrcClassesAreAbstractOrFinal()
+    {
+        $this->assertClassesAreAbstractOrFinal(__DIR__ . '/../../src');
+    }
+
+    public function testSrcClassesHaveTests()
+    {
+        $this->assertClassesHaveTests(
+            __DIR__ . '/../../src',
+            'Localheinz\\Specification\\',
+            'Localheinz\\Specification\\Test\\Unit'
+        );
+    }
+}
